Implement new smoothing strategy at the start of data series
Is used to smooth out pace spikes at the start of an activity, for example.

Series with FORCE-ZERO-START set to #t will use this algorithm.  The name of
the parameter no longer reflects what it does :-).

How it works: An average value is computed for the start of the series and the
average is combined with a few samples at the start of the series.  The number
of samples is controlled by two parameters.