bug fix import and display activities without a distance series
when there is no "dst" series, no "timer" series is created either and the
code did not handle this case cleanly.

* rkt/data-frame.rkt (data-series%/get-index): don't return #f
for the first item in the series.
(df-describe): handle gracefully the case when there are no
statistics.
(df-statistics): fix contract, the function can also return #f

* rkt/inspect-graphs.rkt (graph-view%prepare-render-tree): be
carefull of creating factors for #f series

* rkt/session-df.rkt (make-session-data-frame): set the weight
series only to an existing series
(extract-data): stop-points might be #f

* rkt/time-in-zone.rkt (decoupling): handle df-statistics
returning #f
(update-aerobic-decoupling-for-session)
(update-aerobic-decoupling-for-lap): handle #f adec value