Add a shaded "bests" line to the best avg plot
A shaded area with the best vaue for each duration from a time range can be
added to the best avg plot.  Database version updated to 19.

* inspect-best-avg.rkt: show a configurable "bests" shaded plot
for the best avg data.

* toplevel.rkt (toplevel-window%/on-toplevel-close): call
`clear-metrics-cache' when closing the database.

* metrics.rkt: new file, calculate best avg aggregate metrics,
with a caching system in place for speeding things up.

* dbapp.rkt (schema-version): updated to version 19, now just a
plain function instead of a parameter

* sql/p19-metrics.sql: new file, update a database from v18 to v19

* sql/db-schema.sql (BAVG_CACHE): new table
(SCHEMA_VERSION): updated to 19

* dbapp.rkt (schema-version): updated to version 19, now just a
plain function instead of a parameter