Some small bugfixes
* rkt/inspect-quadrant.rkt
(quadrant-plot-panel%/on-threshold-cadence): be careful with an
'emtpy cadence.
(quadrant-plot-panel%/on-outlier-percentile): if a data series is
already available compute the percentiles without refreshing the
entire plot.

* rkt/widgets.rkt (validating-input-field%/validate): schedule the
timer regardless if the new value is the same as the old.
(validating-input-field%/cb-timer): timer callback will only
notify the user if the new value is different fromt he old.