Cache weather data only while the application is running ... plus some small fixes and improvements
The weather data cache is no longer stored on disk, as it was not well
implemented: after a while the application would just reuse old weather
stations, not picking up new ones.

* rkt/edit-session-weather.rkt
(weather-edit%/on-wstation-selected): display a busy cursor while
fetching weather data

* rkt/weather.rkt (update-session-weather): store just the weather
startion name (the "pws:" prefix, if any, is already part of it)
(wu-cache-db): use a function to get access to the cache data
base, and only open an in-memory database, which caches data for
the current session only.