Large assortment of small fixes and improvements
* rkt/data-frame.rkt (df-raise): actually raise the exception
(data-series%/check-consistency, data-frame%/get-series)
(data-frame%/add-series): clarify exception message
(merge-lists, get-histogram-buckets normalize-histogram): export
these functions

* rkt/fmt-util.rkt (m/s->sec/km, m/s->sec/mi, m/s->sec/100yd)
(m/s->sec/100m): simplify expresions to avoid to many division
operations.

* rkt/intervals.rkt (make-split-intervals, make-climb-intervals):
handle gracefully the case where the data frame does not contain
the required series

* rkt/metrics.rkt (expand-histogram): add buckets that have no
elements so the histograms are even, like `df-histogram` does.

* rkt/series-meta.rkt (axis-pace, axis-stride): clarify reasoning
for bucket slot size, also change stride bucket slot size to
0.001, from 0.01

* rkt/sport-charms.rkt (get-athlete-gender, get-athlete-dob)
(get-athlete-height): update contracts to be more permissive

* rkt/toplevel.rkt (toplevel-window%/run): pop up a notification
if there are no activities in the database

* rkt/trends-bw.rkt (bw-trends-chart%): handle the case when there
is no data to plot.

* rkt/trends-chart.rkt (get-true-min-start-date): handle the case
when there are no activities in the database and the date cannot
be determined.  Also updated contract to allow negative
timestamps.
(generate-timestamps, start-of-day, str->date): update contracts
to allow negative timestamps.

* rkt/trends-hist.rkt (make-render-tree): handle the case when
there is no data to plot.
(hist-trends-chart%/export-data-as-csv): implement data export.

* rkt/trends-pmc.rkt (pmc-trends-chart%/maybe-build-pmc-data):
handle the case when there is no data to plot.

* rkt/trends-tiz.rkt (tiz-trends-chart%): handle the case when
there is not data to plot.

* rkt/trends-trivol.rkt (trivol-trends-chart%): handle the case
when there is not data to plot.

* rkt/trends-tt.rkt (tt-trends-chart%): handle the case when there
is not data to plot.

* rkt/trends-vol.rkt (vol-trends-chart%): handle the case when
there is not data to plot.

* rkt/widgets.rkt (validating-input-field%): add an allow-empty?
field to indicate whether empty fields are valid or not.