Critical Power estimate for pace and power series
The best-avg trend plots can be configured to add a CP estimate line to the
plot.  This works for pace and power data series.  The Critical Power (CP or
CV) and Anaerobic Work Capacity (W' or D') are also displayed on the plots
together with best power and pace estimate for predefined distances.