generate-best-avg-durations -- use a max growth of 300 seconds
The incrementing of best-avg durations (used to determine best avg points) is
limited by a MAX-GROWTH parameter.  With the default growth factor of 1.05,
this effectively kicks in at durations of 6000 seconds or longer (1h 40min).
This is done to keep the precision of the spline interpolation at longer
intervals more accurate.