New sport zone editor
Database schema updated to version 24, to fix a bug in the V_SPORT_ZONE view.

The sport zones in the database have a timestamp, allowing different zones to
be applied to activities (e.g. when heart rate zones are updated, previous
activities still use the old heart rate zones).  This was supported in the
database schema and the rest of the application, but the sport zone editor
always created a "latest" sport zone which applied only to activities that
occurred after the zones were entered.

The new sport zone editor allows editing all sport zones, including creating
and updating previous sport zones.

When zones are updated, the TIME_IN_ZONE data is not updated, and
"Tools/Update Metrics..." needs to be run.  This operation does more work than
necessary and it could be improved in the future.