Refactoring of plot view, some speed improvements
The state of a plot (axis, selected lap, etc) is encapsulated in a PS struct,
and the data for the plot in a PD struct.  These structures are never updated,
instead, new ones are created each time a setting changes.  `update-plot-data`
will than create anew PD structure by reusing things from the previous PD
struct and only recomputing the minimum set needed.

The graph-panel% object queues a new update request on a separate thread each
time something changes and updates the on the canvas with the resulting image.