Synchronize data between views
Changes to sessions and athlete metrics are now synchronized between the
various views, so they don't need to be manually refreshed when things change.
For example, editing the session headline in the session view, will
automatically update the headline in the activity list and calendar view.

Currently, reports and trends are not updated via this mechanism and they need
to be manually refreshed with Ctrl-R.