Filter by labels and equipment for BAVG, HISTOGRAM and SCATTER trend charts
The settings dialogs for the BAVG, HISTOGRAM and SCATTER trend charts have
been updated to allow filtering sessions by labels and equipment in addition
to sport and date.

* rkt/al-widgets.rkt (dbtag-input-field%{refresh-available-tags,
set-contents-from-tag-ids}) new methods

* rkt/metrics.rkt (fetch-candidate-sessions): add #:label-ids and
#:equipment-ids options.

* rkt/trends-bavg.rkt (bavg-params): add labels and equipment
slots.
(bavg-chart-settings%): use a `session-filter%' object to filter
sessions, keep track of labels and equipment in the settings.
(candidate-sessions): updated to filter on labels and equipment.

* rkt/trends-chart.rkt (session-filter%): new widget to filter
sessions on sport, date equipment and labels.  Can be reused by
all trend chart settings dialogs.

* rkt/trends-hist.rkt (hist-params): add labels and equipment
slots.
(hist-chart-settings%): use a `session-filter%' object to filter
sessions, keep track of labels and equipment in the settings.
(insert-plot-snip): use `time-ticks` if time is displayed on the
histogram.
(candidate-sessions): updated to filter on labels and equipment.

* rkt/trends-scatter.rkt (scatter-params): add labels and
equipment slots.
(scatter-chart-settings%): use a `session-filter%' object to filter
sessions, keep track of labels and equipment in the settings.
(candidate-sessions): updated to filter on labels and equipment.