remove 'smart series-meta%/plot-color return value
This was introduced to indicate that swim pace timing plots should be colored
by swim-stroke, but it caused too much trouble in all other plots.
`plot-color` now always returns a valid color.

There's a `series-meta%/plot-color-by-swim-stroke?` method which is checked by
`plot-data-renderer-tree`, which is the only function that cares about it.