Plot Critical Power curve on session Best-Avg plot, fixes #13
... also display a snip with the best efforts for the activity and the
theoretical best efforts from the CP curve.

The Critical Power parameters used are the ones that are applicable to the
session based on the V_CRITICAL_POWER_FOR_SESSION view, and are set in the
"Athlete/Edit Critical Power..." dialog box.

The `edit-cp-dialog%` has been updated to create/edit CP parameters for
swimming, this allows the CP curve to be displayed for swimming activities as
well.

The heat map for the best avg trends plot has been changed to limit the value
of the curve to 100% -- spline interpolation caused it to go above 100% when
two adjacent points had oposing tangents.