Implement Grade Adjusted Pace for running sessions (#15)
Grade Adjusted Pace is an adjustment to the Pace series taking the slope of
the terrain into account.  The new GAP series can be analyzed in all the
session views and trend charts.  Also W'Bal for running is computed using the
GAP series to model more accurately the W' depletion while running uphill.