Update activity list and reports when athlete metrics change #22
Changes to athlete metrics will also trigger 'session-updated events for the
sessions that might be affected by the changes (currently only the *activity
list* view shows body weight for a session).

Reports will also refresh if *any* sessions or athlete metrics have changed --
this is not as find grained as it could be, but it can be difficult to
determine if a change to a session has an effect on the report, and refreshing
reports is relatively fast.