edit-lap-swim: fix bad joining of drill sections
When editing a lap swim activity and trying to delete or split a drill length,
an error would be signaled.  No data was lost, since the entire thing ran as
part of a SQLite transation, but the edit would fail.

* edit-lap-swim.rkt (split-length/rebalance) (join-lengths/rebalance): fix
several silly bugs: use `query-list` instead of `query-value`, use correct
numer of lengths in the `for` loop iteration, and return the lap ID from the
function -- these functions were not exercised when the code was originally
written.

(fixup-swim-session): when rebalancing laps, mark the lap as modified, so that
the summary values are updated for them.