Calculate the quantile bounds using data displayed in the scatter plot
The data in scatter and quadrant plots are filtered out to remove invalid
values, but the quantiles were calculated from the original data from the data
frame.  This resulted in differences between the data bounds and the quantile
bounds.  For example, data points with cadence of 0 and stride of #f would be
filtered out from the data, but the 0 cadence values would be considered for
the quantile bounds.

The quantile bounds are now generated from the same data, omitting invalid
values, making them consistent with the data bounds.