Track the map location when hovering over the elevation plot
A "current" location concept was added to the map-widget%, as well as the
option to auto-pan the map to make the current location in the center.

as part of this, a new data series, map-point is added to the data frame.
This is omitted from the export to CSV, as it contains structures and it is
not of general use outside the application, but it does speed up the location
to GPS position lookups.