Bugfix -- loss of trend charts
Trend charts could be lost if they don't finish loading when the application
shuts down -- this can happen when there are a lot of charts which loaded in
the background when the "Trends" tab was selected.

The new implementation will create the trend charts lazily, but restore data
will be saved correctly even if the charts are not created (due to not being
visited).