Updates and improvements to map-widget%
These are updates made while releasing map-widget% as a separate package.
ActivityLog2 does not use that package yet, but the improvements have been
added to this code base:

Renmed some map-widget% methods:

       `delete-track-group` => `delete-group`,
       `set-track-group-pen` => `set-group-pen`,
       `set-track-group-zorder` => `set-group-zorder`
       `get-bbox` => `get-bounding-box`

Pens and zorder are now maintained separately from the actual tracks, this
means that you can add a track to a group after a call to `set-group-pen` and
the track will use the correct pen for drawing.  Previously,
`set-track-group-pen` updated existing tracks only, so a new track in the same
group would use the default pen.  Similar for the Z order.

The map legend lines are drawn in both metric and imperial, so the map-widget%
does not need to know the preferred measurement system (drop dependency on
"fmt-util.rkt")

`get-bounding-box` now considers any map markers if there are no tracks --
this is so that `center-map` works if only markers are added to the map.

Tile downloader was updated to check that the downloaded tile is actually an
image and not an error reply from the server.  If it is an error reply, mark
it as a "problem tile", this will cause it to be moved to the back of the
download backlog when it is requested again, this is done so that the
downloader does not get stuck on a tile that cannot be downloaded.