don't print an error message when build-id.txt is missing (fixes #26)
The build-id.txt file is generated when the code is built using `racket
build.rkt` and it is only used to identify a build in the "About" dialog.  If
the code is built or run without running build.rkt first, it will print an
error message about the missing build-id.txt file, which is confusing and not
really needed, as the application works fine even if the file is missing.