reduce flicker when the session graphs tab is opened for the first time
The graph-view% objects are created with a 'deleted style, so they don't show
up in the parent panel until they are explicitly added.

Also attempted to reduce the number of time the plots are redrawn, but this
was unsuccessful: attempted to reflow the container so the canvas size is
known when the plot is added, and the `update-snip` will not resize the snip
if the size does not change.