Adjust map tile cache threshold based on canvas size #29
The default cache threshold of 100 was not adequate for large displays where
more than 100 tiles were required to draw the map.  In that case the cache was
rotated in the middle of the draw causing empty bitmaps to be drawn.

The new method allows the map widget to specify how many tiles to keep in the
cache, based on the size of the canvas.  There is also some logging, to help
diagnose issues in the future.