find-meta-for-series can now find the "gap" series
The Grade Adjusted Pace series, "gap", was missing from `all-series-meta`, so
`find-meta-for-series` could not find it.  This caused the Histogram Trend
chart for the gap series to fail.

Also added a contract to `find-meta-for-series` so it fails if it returns #f,
and a similar problem will be easier to identify in the future.