Assorted fixes for the data-frame implementation
* rkt/intervals.rkt (find-climbs): use `in-data-frame` to avoid
creating an intermediate vector while processing the points.

* rkt/data-frame/csv.rkt (df-read/csv): add a #:na option which
allows specifying a string that is the "not available" value in
the CSV.  Some CSV exporters like to use "-" for the "empty" cell.

* rkt/data-frame/df.rkt (df-add-derived): fix typo in variable
name.

* rkt/data-frame/histogram.rkt (samples->buckets): allow handling
samples that are not numbers
(buckets->histogram): allow handling samples that are not numbers.
(df-histogram, histogram-renderer): fix contract

* rkt/data-frame/statistics.rkt (df-statistics): fix contract.