Assorted fixes of bugs found while testing odd FIT files
Wrote up a test for various odd FIT files I collected from the Net over the
years and this revealed many bugs in ActivityLog2.  The list below details
what was fixed.

* .travis.yml: tidied up, moved travis specific scripts in the
.travis folder, added "fit-test.rkt" test run.

* .travis/download-test-data.sh: moved into .travis folder and
updated the files that it downloads.

* rkt/elevation-correction.rkt (lat-lon->tile-code): use
`exact->inexact` to handle cases when `lat` and `lon` are not
floats.

* rkt/series-meta.rkt (axis-elevation, axis-corrected-elevation):
`missing-value` returns `#f`, so missing elevation data points are
not replaced with anything.

* rkt/session-df.rkt (fetch-trackpoins/swim): fix SQL query to
return 0 for duration, if a NULL is stored in the database.
(add-grade-series-1): don't add an empty grade series to the data
frame.
(fixup-pp-series): delete the series if it becomes empty after it
is processed.
(fixup-lrbal-series): delete the series if it becomes empty after
fixing the invalid values.
(add-timer-series): change the way stop points are detected: an
average recording interval is kept while generating the timer
series and any points that is larger than three times the average
is considered a stop point.  "timer" series generation does not
depend on "dst" series anymore.
(extract-data): when the data set is small, stop points are added
as additional points to the data, otherwise, existing points at
the stop are changed to 0.  The `rdp-simplify` is not called on
small data sets, also stop points are explicitely kept in the
dataset using `#:keep-positions`.

* rkt/data-frame/df.rkt (df-get-property): fix contract to allow a
non-procedure to be specified for the fallback when the property
is not found.

* rkt/data-frame/rdp-simplify.rkt (rdp-simplify): add a
`#:keep-positions` argument, this specifies a list of positions
that are not removed and will be present in the output set.

* rkt/data-frame/spline.rkt (spline): raise-argument-error checks
for at least 3 elements in the input data; in the search look use
`=` and `<=` instead of `eqv?` and `<` to find points that are at
the knots.

* rkt/dialogs/activity-edit.rkt
(activity-operations-menu%/on-export-csv): fix argument order to
`df-write/csv` call.

* rkt/fit-file/fit-file.rkt (decode-record-header): fix end bit
position in call to `bitwise-bit-field`; fix syntax mistakes when
decoding developer fields.
(read-fit-records/read-next-record): use 255 when
`read-next-value` returns `#f` (invalid value) -- there are no
invalid values when decoding the header.
(fit-event-dispatcher%/update-timestamp): fix bug in compressed
timestamp handling, previously it was completely incorect.  Also
handle cases where records have a timestamp of 0 (which is
invalid, but FIT files with such records exist).
(activity-builder%/on-record): merge adjacent records if they have
the same timestamp.
(activity-builder%/on-lap): when there is no length record, and we
have to create our own, use all the fields from the lap record in
the new length record.
(activity-builder%/collect-activity): handle session records with
incorrect start-time/timestamp (some cases anyway).

* rkt/session-inspector/inspect-best-avg.rkt
(mean-max-plot-panel%/save-visual-layout)
(mean-max-plot-panel%/set-session): handle the case when the data
frame has no series to plot

* rkt/session-inspector/inspect-graphs.rkt (split-by-factor): do
not loose points when spliting.
(plot-data-renderer-tree): make sure the data frame contains
"swim_stroke" before trying to use it.
(graph-view%/refresh-plot, graph-view%/refresh-cached-bitmap):
handle the case when there is no data to plot.
(graph-panel%/set-session): only use X axis that actually exist in
the data frame

* rkt/session-inspector/inspect-histogram.rkt
(histogram-plot-panel%/save-visual-layout)
(histogram-plot-panel%/set-session): handle the case when the data
frame has no series to plot

* rkt/session-inspector/inspect-map.rkt (map-panel%/set-session):
account for laps that have no data points

* rkt/widgets/notification-banner.rkt
(notification-banner%/on-timer): display how many more messages
are waiting to be displayed.

* rkt/widgets/tab-selector.rkt (tab-selector%/set-selection)
(tab-selector%/select): fix off by one error.

* test/test-util.rkt: new file, containing common utilities from
db-test.rkt

* test/df-test.rkt (rdp-simplify-tests): test `#:keep-positions`

* test/fit-test.rkt: new file, tests various off FIT files that
caused ActivityLog2 to misbehave.