Remove plot-hack.rkt, replace plot-snip/hack with plot-to-canvas
plot-hack.rkt was introduced around Racket 6.2 release to address performance
issues between the plot package (written in Typed Racket) and the GUI package
(written in plain Racket).  As of Racket 6.12, this code is no longer
necessary.

`plot-snip/hack` was replaced with `plot-to-canvas`, as some of the
functionality it provided is still needed (e.g. creating the snips with the
size of the canvas).

Also updated the check-requires.sh script to avoid checking deleted files.