Refactor trend charts to only use hashes for their settings
The interface required by the trend charts configuration data was too
complicated.  It is now replaced by a simpler one, which is also defined in an
interface: `chart-settings-interface<%>`.  All trend charts now store their
configuration data as hash objects, and these hashes are also used for the
actual charting (previously structures were holding this data).

As a result of this change, some trend charts will not be compatible with old
ones defined by the user, and those will have to be deleted and re-created.
There is no migration path defined.

Oh, and actually remove the plot-hack.rkt file...