Create database-opened events and clear caches as needed
* rkt/database.rkt: monitor 'database-opened events and clear the
sid-timestamp-cache

* rkt/dbapp.rkt (add-db-open-callback, del-db-open-callback):
removed.
(set-current-database): log a 'database-opened event

* rkt/metrics.rkt (process-events): process events like
'database-opened and 'session-updated-data and clear the cache as
needed.  This is called by all interface functions to ensure the
cache stays consistent.

* rkt/session-df.rkt (process-events): process events like
'database-opened and 'session-updated-data and clear the cache as
needed.  This is called by all interface functions to ensure the
cache stays consistent.

* rkt/sport-charms.rkt (maybe-init-sport-charms): monitor
'database-opened events and re-initialize.

* rkt/toplevel.rkt (toplevel-window%/on-toplevel-close): no need
to clear caches anymore.

* rkt/dialogs/activity-edit.rkt: no need to call
`clear-session-df-cache` anymore

* test/test-util.rkt (with-fresh-database, with-database): no need
to call `clear-session-df-cache` anymore