Assorted list of small bug fixes
* rkt/bavg-util.rkt (get-aggregate-mmax-heat-map): refine contract
for percentage argument to be between 0 and 1.

* rkt/metrics.rkt (aggregate-mmax-heat-map): fix spelling mistake
in comparison, plus refine contract for percentage argument to be
between 0 and 1.

* rkt/trend-charts/trends-bavg.rkt (mmax-chart-settings%
get-chart-settings): convert heart-percent from 0..100 to 0..1
(mmax-chart-settings% put-chart-settings): convert heart-percent
from 0..1 to 0..100

* rkt/session-df/session-df.rkt (extract-data): don't run
`rdp-simplify` when epsilon is zero.

* rkt/session-inspector/inspect-graphs.rkt (ivl-extents): use <=
as the compare function for `bsearch`, to find the first element,
not the last one when there are multiple ones with the same index.

* rkt/trend-charts/trends-tiz.rkt (tiz-trends-chart%
maybe-fetch-data): adjust start range to not go earlier than the
first activity in the database.

* rkt/trend-charts/trends-trivol.rkt (trivol-trends-chart%
maybe-fetch-data): adjust start range to not go earlier than the
first activity in the database.

* rkt/trend-charts/trends-vol.rkt (vol-trends-chart%
maybe-fetch-data): adjust start range to not go earlier than the
first activity in the database.

* test/trends-test.rkt (check-csv-file): check if too many rows
were added to the exported CSV file in an attempt to catch if
there is a mistake in adjusting the start date for the trends.