Add support for named sport zones
Each sport zone now can have a name defined, and this name is used in all
places where sport zones are used (e.g. as values for histogram plots and as
hover tags for plots)

The series-metadata%/value-formatter was also updated to accept a SPORT and
SID arguments so it can produce formatters that are specific to a sport or
session.