Clear the plot snip in graphs when there is no data to plot
The session inspector graphs are linked together for displaying hover
information.  For each graph, the hover callback runs if there is a plot snip
(`the-plot-snip` is not `#f`), but this variable was not cleared anywhere, so
when a data series, such as "calt", did not exist, the hover callback would
run and `df-get-series` would report an error on the console.