Implement Heatmaps as a trend chart [#11]
Heatmaps display a set of GPS activities overlayed over a map.  They are
currently implemented by drawing all the activities on top of each other,
which is not how heatmaps are usually implemented, but this is the simplest
approach.  The chart can be used to visualize, for example, the training
routes in a season.

* sql/queries/gpspoints.sql: new file

* rkt/trend-charts/trends-heatmap.rkt: new file

* rkt/trend-charts/view-trends.rkt (chart-types): register the trends heatmap
  chart type.

* test/trends-test.rkt (trend-charts-test-suite): add trends heatmap tests

* rkt/widgets/map-widget/map-impl.rkt (map-impl% draw-map-tiles): only request
  a refresh if the `allow-tile-download` is #t, this is done to make the
  trent-chart tests pass, but likely incorrect.

* rkt/plot-util.rkt (move-snip-to): check if the admin and editor exist, in
  case we are asked to move snips which are not inside visible pasteboards.