Replace `snip-canvas%` with `plot-container%`
Added `plot-container` as a submodule dependency and update the source files
to use it.  Removed "rkt/plot-util.rkt".

The `snip-canvas%` class, which used to hold plot snips in the GUI, is
replaced with the `plot-container%` class from the plot-container package.
Same functionality is provided, in addition to this, `plot-container` supports
multiple plots in a single canvas -- dual scatter plots now use this feature
instead of creating a second canvas.

Several plot helper methods are also part of the plot-container package and
they were used directly.