Reporting test results to Azure DevOps
Implemented a custom test runner which runs a rackunit test suite and writes
results to an XML file in JUnit format.  Updated all tests to use this test
runner and updated the azure build pipeline to upload the test results to
Azure DevOps for analysis and reporting.

Individual tests were updated to (1) use the custom test runner, (2) reduce
the verbosity, so less output is logged by default, (3) construct a separate
test case for each individual test and (4) report tests as skipped if they
cannot run (e.g when test data is not available).

The Racket version for the build has been updated to 7.5.