Fix bad XDATA decoding and speed up reading sessions with XDATA
The device that recorded f0029 re-used the developer index from one
application to another -- this is valid for the FIT file format, but the FIT
reader used the application index to link fields to applications, assuming
that it is unique.  Updated the code to use the 128 bit application ID
instead.

Fetch batches of 1000 rows in read-xdata-series to speed up reading sessions
with lots of rows, which contain XDATA series.

Also add Xert application field (wprime_bal) to xdata-defs.json.