Assorted list of small fixes
* rkt/view-activities.rkt (view-activities% get-selected-sid)
  (view-activities% get-selected-sport)
  (view-activities% get-selected-guid): use `and` instead of `when`
  so the methods return #f instead of (void) when no data is
  available.

* rkt/view-last-import.rkt (import-view% get-selected-sid)
(imoport-view% get-selected-sport, import-view% get-selected-guid)
(import-view% after-update): use `and` instead of `when` so the
methods return #f instead of (void) when no data is available.  :

* rkt/session-inspector/inspect-graphs.rkt (ivl-extents): handle
the case when END is outside the timestamp range.
(graph-view% cached-bitmap-token): initialize it to -1 to allow
`=` to work on the initial value.
(graph-view% refresh-plot): avoid unnecessary refresh call.
(graph-view% refresh): only call `refresh-cached-bitmap` if there
is a bitmap (plot snip) already.

* rkt/session-inspector/inspect-overview.rkt
(*run-summary-fields*, *bike-summary-fields*,
*swim-summary-fields*, *other-summary-fields*): use two digits of
precision for the intensity field.

* rkt/widgets/grid-pane.rkt (grid-pane% child-placement): ensure
the child width and height are also exact integers.

* rkt/widgets/map-widget/map-snip.rkt (map-snip%/c): resize-to-fit
also accepts numbers as group indentifiers

* rkt/widgets/map-widget/map-widget.rkt (map-widget%/c):
resize-to-fit also accepts numbers as group indentifiers