Hover label improvements on various plots
Hover labels are more opaque and bigger, making them easier to read (updated
plot-container package)

Hover labels on plots have been updated to show units for the displayed values
(only for some of the series).  To help with this,
series-metadata%/value-formatter was updated to produce formatters which
append the units to the values -- this is an optional argument, as formatters
are used for plot-ticks as well.

On histogram plots, the hover labels will also display the X axis value for
each bar.

On time/distance charts, the hover labels are displayed at the bottom of the
chart, to avoid the labels jumping up and down on the various plots.  Also,
the code which displays the labels, `find-y-values`, has been updated to avoid
lookups during the mouse callback: `make-y-values-finder` will do the lookups
in advance and create a closure for the mouse callback.