Avoid unnecessary refresh calls in map-widget%
It looks like the paint call for a canvas% comes in before the show
notification, so update the zoom level in the paint method as well, if
delayed-resize-to-fit is set.

Also avoid refresh when the map implementation is notified of a size change --
the refresh is handled by the widget or snip classes.