DB migration 35, add PMAX to CRITICAL_POWER table
The 3 Parameter Critical Power model supports a "max power" value, add this to
the CRITICAL_POWER table and related views (if this is NULL, the data in the
row is for a CP2 model).

NOTE: technically, the CP3 model uses a "time delay", called `k`, in addition
to the CP and WPRIME values, but its value is unintuitive to the user, and it
can be easily deduced from PMAX, CP and WPRIME, making PMAX a better choice
for storage.