Grade series calculation improvements (avoid spikes)
We won't compute the grade for very short segments, instead these segments are
set aside in the main pass and interpolated later from the grade of adjacent
segments.  This avoids grade spikes in the activities that I tested with...

The result of this change is that the Open Water Swim (OWS) portion of
f0029.fit no longer has a grade series.  This is probably OK, as there is no
real concept of elevation or grade for OWS, but might hide problems somewhere
else.  For now, this update produces adequate results for the activities I
care about.