Use the `map-widget` package (AB#10)
The map-widget% and releated utilities from "rkt/widgets/map-widget" have been
replaced with the `map-widget` package which is added as a submodule to the
"pkgs/" folder.  Some function names have been updated to match the names used
by the `map-widget` package.

Also created the `the-application` package, which stores the `get-pref`,
`put-pref` and `data-directory` functions, so they can be used by the
`map-widget` package to store its data file with the rest of the ActivityLog2
data files.

Build documentation was also updated.