Reorganize the build pipeline
Tests are only run on Linux.  The Windows build takes significantly longer to
run the tests and so far there was no test that failed on one platform only.

The tests for the dependent packages in the "pkgs/" folder are also run --
this ensures that the packages are compatible with each other in the selected
version.