Altitude functions which didn't make it
As part of #51, several improvements were made to the elevation correction
algorithm.  This commit contains the smoothing algorithms which did not make
it into the final version:

* `smooth-altitude/friedl`: is the smoothing algorithm used previously.  It is
a variation of RDP simplification

* `smooth-altitude/rdp`: is a RDP based simplification algorithm

* `instant-elevation-change`: is an algorithm used to determine how much the
elevation changes from one point to the next -- it can be used to calculate
epsilon values for the two smooth-altitude algorithms, instead of using fixed
numbers, since it appears that fixed values are not suitable for all
activities and terrain types.