Clarify import activities functionality (#54)
Renamed "Import from directory..." to "Import new activities..." and "Import
activity..." to "Import single activity..." and added the Ctrl+Shift+I
shortcut for the latter.

The last directory from which single activities are imported is now
remembered, and the file dialog is opened again in the last location it was
imported from.