Update fit readed to better recognize session records
Unlike length and laps, session records in FIT files can appear anywhere in
the FIT file, and it is not always clear which laps belong to which sessions,
especially when multiple sessions are present (e.g. multisport activities).
Different devices generate session records with inconsistent timestamps and/or
durations.  Added f0043.fit to the test cases and other test cases also pass.