Fix Y plot range for session MMAX plot
When a session had critical power (CP) data, the maximum critical power value
was considered when determining the range of the plot, even if the series was
not a power series, resulting in a bad Y plot range.

CP is now only considered if the series has a CP estimate (`have-cp-estimate?`
returns #t).  This is still not perfect, see AB#33, but works in most cases.

Also fixed series name for the FTHR analisys dialog.