Assorted small fixes
* docs/README.md: document how to trace module loads

* etc/esc-demo.rkt (esc-control-demo-snip% draw): save and restore
smoothing setting on the drawing context, see #59

* rkt/dialogs/edit-session-tss.rkt (edit-session-tss-dialog%):
replace "Threshold Power" with FTP -- the term appears to be free
to use.

* rkt/models/fthr.rkt (best-segment): return #f when the duration
is longer than the available data in the data frame, but throw an
exception if the series is not present (also updated contract).

* rkt/session-df/native-series.rkt (axis-hr-zone value-formatter)
(axis-power-zone value-formatter): do somethig sensible when sport
zones are not available.

* rkt/session-df/session-df.rkt (make-plot-renderer/swim-stroke):
new argument: Y-RANGE defining the low and high range of the plot.

* rkt/session-inspector/inspect-graphs.rkt
(plot-data-renderer-tree): use the Y-RANGE when rendering a swim
plot colorized by swim strokes.

* rkt/widgets/readme.md: remove references to map-widget%, as this
is now a separate package.

* test/fthr-test.rkt (fthr-test-suite): add tests for updated
`best-segment` behavior.