Add tool to clear power spikes (AB#31)
The tool allows clearing power spikes recorded from a cycling activity and it
is available from the "Activity/Clear Power Spikes..." menu.  This is a port
of the "etc/power-spikes.rkt" tool, which was accessible only from the Racket
REPL.

As part of this, some common code was refactored from the fthr-analysis.rkt
file, and some fixes to fthr-analysis.rkt were also done to deal gracefully
with activities that have incomplete data.

Also as part of this, a new `box-and-whiskers` plot type has been created, see
"rkt/widgets/box-and-whiskers.rkt"