Update data-frame package and improve GPX export
data-frame renamed some of the functions which mutate the data-frame object to
include "!" at the end of the name.  Updated all functions to the new names.

Also added handling for the case when GPX data cannot be exported from an
activity (e.g. because there is no latitude/longitude data), as well as tests
for the GPX export functionality.  The data-frame changes themselves also
include imprivements to GPX export, in particular, more data series can be
exported (HR, Cadence, Speed, Power and Distance)