Mark activities with labels on trend charts (AB#36)
For some trend charts, the user can choose session labels to show on the chart
and timestamps of the sessions containing the selected labels will be marked
on the plot.  Charts supporting this feature are the body weight, aerobic
decoupling, PMC, time in zone, training volume and triathlon training volume.

This allows, for example, to mark races on the performance management chart,
by marking the race sesssion with th "race" label and selecting the "race"
label on the PMC chart.  Given that "note" activities with future dates can be
created and assigned labels, this feature also allows marking future races or
other evens on the chart.