Replace `db-get-last-pk` with `db-insert` (AB#38)
The new `db-insert` function will run an INSERT SQL statement and return the
ROWID for the inserted row, this is based on the Racket db library returning a
`simple-result` with an `insert-id` key for inserts.

The old `db-get-last-pk` was removed and all call pattens of `query-exec` +
`get last PK` were replaced with a single `db-insert` call.