Handle gracefully deleted activities in Mean-Max window (#70)
The Mean-Max session inspector window will retrieve the activity start times
to display it on hover when the "best 6 week" line is shown.  However, an
activity might have been deleted between the "bests" data was determined and
when the user hovers over the plot with the mouse over the plot.  Handle this
case gracefully.